Retested all routes, made small corrections to some erroring logic

+++ router/routes/posts.js
@@ -21,18 +21,25 @@ posts
     .post('/add_post', async (req, res, next) => {
         try {
 
-            let newPost = req.body;
+            if (req.body.author !== req.body.username) {
+              
+                let error = new Error("Attempting to create a post where author is not equal to username")
+                error.status = 400;
+                return next(error);
+            }
+                
+            let validationResult = await validatePostBody({ req: req });
 
-            let validationResult = validatePostBody({ requestBody: newPost });
-
-            if (validationResult === '') {
-                let result = await Post.create(newPost);
+            if (validationResult === null) {
+                let result = await Post.create(req.body);
                 return res.status(200).send(result);
             } else {
                 let error = new Error(validationResult);
                 error.status = 400;
                 return next(error);
             };
+
+
 
         } catch (err) {
             let error = new Error(err);
@@ -42,25 +49,33 @@ posts
     })
 
     .patch('/update_post/new_title', async (req, res, next) => {
+        let error; 
+
         try {
 
-            if (!req.query.newTitle) {
-                let error = new Error("Did not include new title");
+            if (!req.query.objectId) {
+                let error = new Error("Update post title missing object ID query");
                 error.status = 400;
-                return error;
+                return next(error);
             };
-    
-            let error = await validatePostBody({ req: req });
-    
+
+            if (!req.query.newTitle) {
+                error = new Error("Did not include new title");
+                error.status = 400;
+                return next(error);
+            };
+
+            error = await validatePostBody({ req: req });
+
             if (!error) {
-                let updated = await Post.updateOne( { _id: req.query.objectId }, { title: req.query.newTitle });
+                let updated = await Post.updateOne({ _id: req.query.objectId }, { title: req.query.newTitle });
                 res.send(updated);
             } else {
                 return next(error);
             };
 
         } catch (err) {
-            let error = new Error(err);
+            error = new Error(err);
             error.status = 400;
             return next(error);
         };
@@ -69,20 +84,26 @@ posts
     .patch('/update_post/new_body', async (req, res, next) => {
         try {
 
+            if (!req.query.objectId) {
+                error = new Error("Update post title missing object ID query");
+                error.status = 400;
+                return error;
+            };
+
             if (!req.query.newBody && !updatingTitle) {
                 let error = new Error("Did not include new body");
                 error.status = 400;
             };
 
-            let error = await validatePostBody({ req: req, updatingTitle: false});
+            let error = await validatePostBody({ req: req, updatingTitle: false });
 
             if (!error) {
-                let updated = await Post.updateOne( { _id: req.query.objectId }, { body: req.query.newBody });
+                let updated = await Post.updateOne({ _id: req.query.objectId }, { body: req.query.newBody });
                 res.send(updated);
             } else {
                 return next(error);
             };
-        
+
         } catch (err) {
             let error = new Error(err);
             error.status = 400;
@@ -93,10 +114,17 @@ posts
 
     .delete('/delete_post', async (req, res, next) => {
         try {
-            let error = await validatePostBody( { req: req } );
+
+            if (!req.query.objectId) {
+                error = new Error("Update post title missing object ID query");
+                error.status = 400;
+                return error;
+            };
+
+            let error = await validatePostBody({ req: req });
 
             if (!error) {
-                let deleted = await Post.findByIdAndDelete( { _id: req.query.objectId });
+                let deleted = await Post.findByIdAndDelete({ _id: req.query.objectId });
                 res.send(deleted);
             } else {
                 return next(error);
@@ -110,7 +138,6 @@ posts
     });
 
 posts.use((err, req, res, next) => {
-    console.log(err);
     res.status(err.status).send(err.message);
 });
